requirements.txt
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.3
scipy==1.9.3
seaborn==0.12.2
scikit-maad==1.4.1
pyyaml>=6.0.1
birdnetlib
tensorflow==2.17.0
